ExaminationImaesCreateWorking

+++ MedicalSystemApp/Controllers/ExaminationImagesController.cs
@@ -2,6 +2,9 @@
 using Microsoft.EntityFrameworkCore;
 using MedicalSystemApp.Models;
 using MedicalSystemApp.Repositories;
+using System.IO;
+using System;
+using System.Threading.Tasks;
 
 namespace MedicalSystemApp.Controllers
 {
@@ -9,7 +12,7 @@ namespace MedicalSystemApp.Controllers
     {
         private readonly RepositoryFactory _factory;
         private readonly IExaminationImageRepository _imageRepo;
-        private readonly IExaminationRepository _examRepo; 
+        private readonly IExaminationRepository _examRepo;
 
         public ExaminationImagesController(RepositoryFactory factory)
         {
@@ -38,11 +41,31 @@ namespace MedicalSystemApp.Controllers
             ViewBag.Examinations = exams;
             return View();
         }
-
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create(ExaminationImage image)
+        public async Task<IActionResult> Create(ExaminationImage image, IFormFile imageFile)
         {
+            if (imageFile == null || imageFile.Length == 0)
+            {
+                ModelState.AddModelError("imageFile", "Please upload an image.");
+                ViewBag.Examinations = await _examRepo.GetAllAsync();
+                return View(image);
+            }
+
+            var uploadsFolder = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads");
+            Directory.CreateDirectory(uploadsFolder);
+
+            var uniqueFileName = Guid.NewGuid().ToString() + "_" + Path.GetFileName(imageFile.FileName);
+            var filePath = Path.Combine(uploadsFolder, uniqueFileName);
+
+            using (var stream = new FileStream(filePath, FileMode.Create))
+            {
+                await imageFile.CopyToAsync(stream);
+            }
+
+            // Save the relative path to the database
+            image.ImagePath = "/uploads/" + uniqueFileName;
+
             if (!ModelState.IsValid)
             {
                 ViewBag.Examinations = await _examRepo.GetAllAsync();
@@ -52,6 +75,8 @@ namespace MedicalSystemApp.Controllers
             await _imageRepo.AddAsync(image);
             return RedirectToAction(nameof(Index));
         }
+
+
 
         public async Task<IActionResult> Edit(int id)
         {
@@ -67,7 +92,7 @@ namespace MedicalSystemApp.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, ExaminationImage image)
+        public async Task<IActionResult> Edit(int id, ExaminationImage image, IFormFile imageFile)
         {
             if (id != image.Id)
                 return BadRequest("Image ID mismatch");
@@ -76,6 +101,26 @@ namespace MedicalSystemApp.Controllers
             {
                 ViewBag.Examinations = await _examRepo.GetAllAsync();
                 return View(image);
+            }
+
+            if (imageFile != null && imageFile.Length > 0)
+            {
+                // Define the uploads folder
+                var uploadsFolder = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads");
+                Directory.CreateDirectory(uploadsFolder);
+
+                // Generate unique file name
+                var uniqueFileName = Guid.NewGuid().ToString() + "_" + Path.GetFileName(imageFile.FileName);
+                var filePath = Path.Combine(uploadsFolder, uniqueFileName);
+
+                // Save the file to the server
+                using (var stream = new FileStream(filePath, FileMode.Create))
+                {
+                    await imageFile.CopyToAsync(stream);
+                }
+
+                // Store relative path to the image in database
+                image.ImagePath = "/uploads/" + uniqueFileName;
             }
 
             try
@@ -105,7 +150,19 @@ namespace MedicalSystemApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            await _imageRepo.DeleteAsync(id);
+            var image = await _imageRepo.GetByIdAsync(id);
+            if (image != null)
+            {
+                var filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", image.ImagePath.TrimStart('/'));
+
+                if (System.IO.File.Exists(filePath))
+                {
+                    System.IO.File.Delete(filePath);
+                }
+
+                await _imageRepo.DeleteAsync(id);
+            }
+
             return RedirectToAction(nameof(Index));
         }
